feat(KeyboardManipulator): Add keyboard manipulator for camera movement
A keyboard manipulator was added for camera movement. The key bindings are re-configurable, but by
default, it uses "WASD" and arrow keys for horizontal movement, and the space bar and shift key for
vertical movement. Movement starts when keys are pressed, and it ends when they are released. The
net movement will be in the average direction of all keys currently pressed. The speed is also
re-configurable.

An example was added of simple terrain exploration.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>