feat(KeyboardCameraManipulator): Update direction when camera is modified
Previously, the direction would not be updated if the camera was modified (by doing things such as
moving changing the direction of view with a mouse manipulator). Now, the direction of the camera
gets updated, so if the camera changes direction during movement, the direction of movement will
also update.