feat(MouseCameraTrackballFirstPersonManipulator): Always orthogonalize viewUp
It was an option previously. However, it seems to cause issues when looking down all the way, and
other parts of the code may be expecting an orthogonalized view up. If we want to change the "up"
direction of the KeyboardCameraManipulator, we should do it in the KeyboardCameraManipulator, rather
than in the MouseCameraTrackballFirstPersonManipulator code.